+++ MonoTools.Debugger/VisualStudio/AD7BoundBreakpoint.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Debugger.Interop;
 
-namespace MonoTools.Debugger.Debugger.VisualStudio
+namespace MonoTools.Debugger.VisualStudio
 {
     internal class AD7BoundBreakpoint : IDebugBoundBreakpoint2, IDebugBreakpointResolution2
     {
