+++ MonoTools.Debugger/VisualStudio/AD7Engine.cs
@@ -5,7 +5,7 @@ using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Debugger.Interop;
 using Microsoft.MIDebugEngine;
 
-namespace MonoTools.Debugger.Debugger.VisualStudio
+namespace MonoTools.Debugger.VisualStudio
 {
     [ComVisible(true)]
     [Guid(AD7Guids.EngineString)]
