+++ MonoTools.Debugger/VisualStudio/AD7ProgramNode.cs
@@ -3,7 +3,7 @@ using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Debugger.Interop;
 using Microsoft.MIDebugEngine;
 
-namespace MonoTools.Debugger.Debugger.VisualStudio
+namespace MonoTools.Debugger.VisualStudio
 {
     public class AD7ProgramNode : IDebugProgramNode2
     {
