+++ MonoTools.Debugger/VisualStudio/MonoProcess.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Debugger.Interop;
 
-namespace MonoTools.Debugger.Debugger.VisualStudio
+namespace MonoTools.Debugger.VisualStudio
 {
     public class MonoProcess : IDebugProcess3
     {
