+++ MonoTools.SharedLib/MonoTools.Debugger.Library.csproj
@@ -77,9 +77,8 @@
     <Compile Include="..\CommonAssemblyInfo.cs">
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="ApplicationType.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Async.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StreamWrapper.cs" />
     <Compile Include="ExtensionMethods.cs" />
     <Compile Include="MonoLogger.cs" />
