+++ MonoTools.SharedLib/Server/MonoWebProcess.cs
@@ -6,12 +6,29 @@ using System.IO;
 using NLog;
 
 namespace MonoTools.Debugger.Library {
-	internal class MonoWebProcess : MonoProcess {
+
+	public class MonoWebProcess : MonoProcess {
 		private static readonly Logger logger = LogManager.GetCurrentClassLogger();
 		public string Url { get; private set; }
 		Frameworks Framework { get; set; } = Frameworks.Net4;
 
 		public MonoWebProcess(Frameworks framework = Frameworks.Net4, string url = null) { Framework = framework; Url = url; }
+
+		public static string SSLXpsArguments() {
+			var a = Assembly.GetExecutingAssembly();
+			var path = Path.GetTempPath();
+			var cer = Path.Combine(path, "MonoTools.CARoot.cer");
+			var pvk = Path.Combine(path, "MonoTools.CARoot.pvk");
+			using (var r = a.GetManifestResourceStream("MonoTools.Debugger.Library.Server.CARoot.cer"))
+			using (var f = new FileStream(cer, FileMode.Create, FileAccess.Write, FileShare.None)) {
+				r.CopyTo(f);
+			}
+			using (var r = a.GetManifestResourceStream("MonoTools.Debugger.Library.Server.CARoot.pvk"))
+			using (var f = new FileStream(pvk, FileMode.Create, FileAccess.Write, FileShare.None)) {
+				r.CopyTo(f);
+			}
+			return $" --https --cert=\"{cer}\" --pkfile=\"{pvk}\" --pkpwd=0192iw0192IW";
+		}
 
 		internal override Process Start(string workingDirectory) {
 			string monoBin = MonoUtils.GetMonoXsp(Framework);
@@ -28,18 +45,7 @@ namespace MonoTools.Debugger.Library {
 				var ssl = uri.Scheme.StartsWith("https");
 				procInfo.Arguments += $" --port={port}";
 				if (ssl) {
-					var a = Assembly.GetExecutingAssembly();
-					var cer = Path.Combine(workingDirectory, "CARoot.cer");
-					var pvk = Path.Combine(workingDirectory, "CARoot.pvk");
-					using (var r = a.GetManifestResourceStream("MonoTools.SharedLib.Server.CARoot.cer"))
-					using (var f = new FileStream(cer, FileMode.Create, FileAccess.Write, FileShare.None)) {
-						r.CopyTo(f);
-					}
-					using (var r = a.GetManifestResourceStream("MonoTools.SharedLib.Server.CARoot.pvk"))
-					using (var f = new FileStream(pvk, FileMode.Create, FileAccess.Write, FileShare.None)) {
-						r.CopyTo(f);
-					}
-					procInfo.Arguments += $" --https --cert=\"{cer}\" --pkfile=\"{pvk}\" --pkpwd=0192iw0192IW";
+					procInfo.Arguments += SSLXpsArguments();
 				}
 			}
