--- MonoTools.VSExtension/VSCommandTable.vsct
@@ -1,74 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<CommandTable xmlns="http://schemas.microsoft.com/VisualStudio/2005-10-18/CommandTable" xmlns:xs="http://www.w3.org/2001/XMLSchema">
-
-  <Extern href="stdidcmd.h"/>
-  <Extern href="vsshlids.h"/>
-
-  <Commands package="guidMonoDebugger_VS2013Pkg">
-    <Menus>
-      <Menu guid="guidMonoDebugger_VS2013CmdSet" id="menuIDMainMenu" priority="0x100"
-        type="Menu">
-        <Parent guid="guidSHLMainMenu" id="IDG_VS_MM_BUILDDEBUGRUN" />
-        <Strings>
-          <ButtonText>Mono</ButtonText>
-          <CommandName>MonoTools</CommandName>
-        </Strings>
-      </Menu>
-    </Menus>
-    
-    <Groups>
-      <Group guid="guidMonoDebugger_VS2013CmdSet" id="TopLevelMenuGroup" priority="0x0600">
-        <Parent guid="guidMonoDebugger_VS2013CmdSet" id="menuIDMainMenu"/>
-      </Group>
-    </Groups>
-    
-    <Buttons>
-      <Button guid="guidMonoDebugger_VS2013CmdSet" id="cmdLocalDebugCode" priority="0x0200" type="Button">
-        <Parent guid="guidMonoDebugger_VS2013CmdSet" id="TopLevelMenuGroup" />
-        <CommandFlag>DynamicVisibility</CommandFlag>
-        <CommandFlag>DefaultInvisible</CommandFlag>
-        <!--<Icon guid="guidImages" id="publish" />-->
-        <Strings>
-          <ButtonText>Debug with Mono (local)</ButtonText>
-        </Strings>
-      </Button>
-
-      <Button guid="guidMonoDebugger_VS2013CmdSet" id="cmdRemodeDebugCode" priority="0x0300" type="Button">
-        <Parent guid="guidMonoDebugger_VS2013CmdSet" id="TopLevelMenuGroup" />
-        <CommandFlag>DynamicVisibility</CommandFlag>
-        <CommandFlag>DefaultInvisible</CommandFlag>
-        <!--<Icon guid="guidImages" id="publish" />-->
-        <Strings>
-          <ButtonText>Debug with Mono (remote)</ButtonText>
-        </Strings>
-      </Button>
-      
-      <Button guid="guidMonoDebugger_VS2013CmdSet" id="cmdOpenLogFile" priority="0x0100" type="Button">
-        <Parent guid="guidMonoDebugger_VS2013CmdSet" id="TopLevelMenuGroup" />
-        <Strings>
-          <ButtonText>Open Logfile</ButtonText>
-        </Strings>
-      </Button>
-    </Buttons>
-    
-    <Bitmaps>
-      <Bitmap guid="guidImages" href="Resources\Images.png" usedList="publish"/>      
-    </Bitmaps> 
-  </Commands>
-  
-  <Symbols>
-    <GuidSymbol name="guidMonoDebugger_VS2013Pkg" value="{27D183E9-5D2B-44D6-9EC8-2DB329096DF7}" />
-    
-    <GuidSymbol name="guidMonoDebugger_VS2013CmdSet" value="{9EF3EF5E-965C-4443-A78A-947849FBA55A}">
-      <IDSymbol name="TopLevelMenuGroup" value="0x1020" />
-      <IDSymbol name="cmdRemodeDebugCode" value="0x0100" />
-      <IDSymbol name="cmdLocalDebugCode" value="0x0101" />
-      <IDSymbol name="menuIDMainMenu" value="0x0102" />
-      <IDSymbol name="cmdOpenLogFile" value="0x0104" />
-    </GuidSymbol>
-    
-    <GuidSymbol name="guidImages" value="{5f91711c-12a6-4a4e-875a-352c3cf839c8}" >
-      <IDSymbol name="publish" value="1" />
-    </GuidSymbol>
-  </Symbols>
-</CommandTable>
