+++ MonoTools.VSExtension/VSPackage.cs
@@ -15,10 +15,8 @@ using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.Shell.Settings;
 using Microsoft.Win32;
 using MonoTools.Debugger.Library;
-using MonoTools.Debugger.Debugger;
-using MonoTools.Debugger.VSExtension.Settings;
-using MonoTools.Debugger.VSExtension.Views;
-using MonoTools.Debugger.VSExtension.MonoClient;
+using MonoTools.Debugger;
+using MonoTools.VSExtension.Settings;
 using NLog;
 using Process = System.Diagnostics.Process;
 using Microsoft.MIDebugEngine;
@@ -33,8 +31,8 @@ namespace MonoTools.VSExtension {
 	[Guid(Guids.MonoToolsPkgString)]
 	public sealed class VSPackage : Package, IDisposable {
 		private static readonly Logger logger = LogManager.GetCurrentClassLogger();
-		private MonoVisualStudioExtension monoExtension;
 		private MonoDebugServer server = new MonoDebugServer();
+		private Services services;
 
 		protected override void Initialize() {
 			var settingsManager = new ShellSettingsManager(this);
@@ -43,10 +41,10 @@ namespace MonoTools.VSExtension {
 			MonoLogger.Setup();
 			base.Initialize();
 			var dte = (DTE)GetService(typeof(DTE));
-			monoExtension = new MonoVisualStudioExtension(dte);
+			services = new Services(dte);
 			TryRegisterAssembly();
 
-			ErrorList.Initialize(this);
+			ErrorsWindow.Initialize(this);
 
 			/* Application.Current.Resources.MergedDictionaries.Add(new ResourceDictionary {
 				Source = new Uri("/MonoTools;component/Resources/Resources.xaml", UriKind.Relative)
@@ -117,9 +115,7 @@ namespace MonoTools.VSExtension {
 		}
 
 		private void OpenLogFile(object sender, EventArgs e) {
-			if (File.Exists(MonoLogger.LoggerPath)) {
-				Process.Start(MonoLogger.LoggerPath);
-			}
+			services.OpenLogFile();
 		}
 
 		private void TryRegisterAssembly() {
@@ -201,116 +197,49 @@ namespace MonoTools.VSExtension {
 
 		private readonly CancellationTokenSource cts = new CancellationTokenSource();
 
-		private async void DebugMonoClicked(object sender, EventArgs e) {
-			var dte = GetService(typeof(DTE)) as DTE;
-			var ex = new MonoVisualStudioExtension(dte);
-			var startup = ex.GetStartupProject();
-			bool isWebApp = ((object[])startup.ExtenderNames).Any(x => x.ToString() == "WebApplication");
-			string host = null;
-			if (isWebApp) {
-				var ext = startup.Extender["WebApplication"];
-			} else {
-				try {
-					if (startup.ConfigurationManager.ActiveConfiguration.Properties.Item("RemoteDebugEnabled")?.Value?.ToString().ToLower()  == "true") {
-						host = startup.ConfigurationManager.ActiveConfiguration.Properties.Item("RemoteDebugMachine")?.Value?.ToString();
-					}
-				} catch { }
-			}
-			/* if (host != null) {
-				Properties monoHelperProperties = dte.Properties["MonoTools", "General"];
-				string ports = (string)monoHelperProperties.Item("MonoDebuggerPorts").Value;
-				int msg, debug, discoveryport;
-				MonoDebugServer.ParsePorts(ports, out msg, out debug, out discoveryport);
-				var discovery = new MonoServerDiscovery(discoveryport);
-				var info = await discovery.SearchServer(host, cts.Token);
-				if (info != null) host = info.IpAddress.ToString();
-				else host = null;
-			} */
-			StartDebug(host);
-		}
-
 		private DTE2 GetDTE() {
 			return GetService(typeof(DTE)) as DTE2;
 		}
 
 		private void StartMonoMenuItemClicked(object sender, EventArgs e) {
-			Services.StartMono(GetDTE());
+			services.Start();
+		}
+		private async void DebugMonoClicked(object sender, EventArgs e) {
+			services.StartDebug();
 		}
 
 		private void XBuildMenuItemClicked(object sender, EventArgs e) {
-			Services.XBuild(GetDTE());
+			services.XBuild();
 		}
 
 		private void XRebuildMenuItemClicked(object sender, EventArgs e) {
-			Services.XBuild(GetDTE(), true);
+			services.XBuild(true);
 		}
 
 		private void XBuildProjectMenuItemClicked(object sender, EventArgs e) {
-			Services.XBuildProject(GetDTE());
+			services.XBuildProject();
 		}
 
 		private void XRebuildProjectMenuItemClicked(object sender, EventArgs e) {
-			Services.XBuildProject(GetDTE(), true);
+			services.XBuildProject(true);
 		}
 
 		private void AddPdb2MdbToProjectMenuItemClicked(object sender, EventArgs e) {
-			Services.AddPdb2MdbToProject(GetDTE());
+			services.AddPdb2MdbToProject();
 		}
 
 		private void MoMAClicked(object sender, EventArgs e) {
-			Services.MoMASolution(GetDTE());
+			services.MoMASolution();
 		}
 
 		private void MoMAProjectClicked(object sender, EventArgs e) {
-			Services.MoMAProject(GetDTE());
+			services.MoMAProject();
 		}
-
-		private async void StartDebug(string host) {
-			try {
-				if (server != null) {
-					server.Stop();
-					server = null;
-				}
-
-				monoExtension.BuildSolution();
-
-				if (host == null) {
-					using (server = new MonoDebugServer(true)) {
-						server.Start();
-						await monoExtension.AttachDebugger(MonoProcess.GetLocalIp().ToString(), true);
-					}
-				} else {
-					await monoExtension.AttachDebugger(host, false);
-				}
-			} catch (Exception ex) {
-				logger.Error(ex);
-				if (server != null) server.Stop();
-				MessageBox.Show(ex.Message, "MonoTools.Debugger", MessageBoxButton.OK, MessageBoxImage.Error);
-			}
-		}
-
-		/*
+		
 		private void DebugRemoteClicked(object sender, EventArgs e) {
-			StartSearching();
+			services.StartSearching();
 		}
-
-		private async void StartSearching() {
-			var dlg = new ServersFound();
-
-			if (dlg.ShowDialog().GetValueOrDefault()) {
-				try {
-					monoExtension.BuildSolution();
-					if (dlg.ViewModel.SelectedServer != null)
-						await monoExtension.AttachDebugger(dlg.ViewModel.SelectedServer.IpAddress.ToString());
-					else if (!string.IsNullOrWhiteSpace(dlg.ViewModel.ManualIp))
-						await monoExtension.AttachDebugger(dlg.ViewModel.ManualIp);
-				} catch (Exception ex) {
-					logger.Error(ex);
-					MessageBox.Show(ex.Message, "MonoTools.Debugger", MessageBoxButton.OK, MessageBoxImage.Error);
-				}
-			}
-		} */
-
+		
 		#region IDisposable Members
 		private bool disposed = false;
 		protected override void Dispose(bool disposing) {
