+++ MonoTools.VSExtension/Views/ServersFound.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="MonoTools.Debugger.VSExtension.Views.ServersFound"
+﻿<Window x:Class="MonoTools.VSExtension.Views.ServersFound"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="Search MonoTools.Debugger Server" Height="199" Width="430" ResizeMode="NoResize"
@@ -18,7 +18,7 @@
                             <RowDefinition Height="Auto" />
                             <RowDefinition Height="Auto" />
                         </Grid.RowDefinitions>
-                        <Image Height="65" Source="Resources/MonoLogo.png" />
+                        <Image Height="65" Source="MonoToolsLogo.png" />
                         <TextBlock Grid.Row="1" Text="{Binding Message}" />
                         <TextBlock Grid.Row="2" Text="{Binding IpAddress}" />
                     </Grid>
